Move some Hono handler to exactly match the original Next physical path
