+++ front-api/routes/w/[wId]/assistant/conversations/[cId]/files/[...rel].ts
